test(channels): cover PM2 runner env
